Rename handle_config module to config
